Extract service snapshot DTOs
